requirements.txt
wheel==0.23.0
vcrpy
requests
python-dateutil
termcolor
tabulate
pytz
tzlocal
mock
